# Conflicts:
#	sdk/include/thinkos.h
#	sdk/include/thinkos/dbgmon.h
#	sdk/include/thinkos/kernel.h
#	sdk/include/thinkos/string.h
#	sdk/include/thinkos/syscalls.h
#	sdk/include/trace.h
#	sdk/libthinkos-dbg/dmgetc.c
#	sdk/libthinkos-dbg/dmgets.c
#	sdk/libthinkos-dbg/dmscanf.c
#	sdk/libthinkos-dbg/thinkos_dbgmon.c
#	sdk/libthinkos-krn/Makefile
#	sdk/libthinkos-krn/misc/exec.c
#	sdk/libthinkos-krn/misc/thread_inf_set.c
#	sdk/libthinkos-krn/thinkos_idle.c
#	sdk/libthinkos-krn/thinkos_svc.c
#	sdk/libthinkos-krn/thinkos_thread.c
#	sdk/libthinkos-krn/thinkos_vector.S
#	sdk/libthinkos/dbgmon/dbgmon_scanf.c
#	sdk/libthinkos/dbgmon/dmgetc.c
#	sdk/libthinkos/dbgmon/dmscanf.c
#	sdk/libthinkos/misc/strlen.c
#	sdk/libthinkos/misc/thread_inf_set.c
#	tools/dcclog.exe

+++ sdk/libthinkos-krn/thinkos_ctl.c
@@ -72,7 +72,7 @@ static int thinkos_cycnt_get(uint32_t cycnt[], unsigned int max)
 	uint32_t cyccnt;
 	int32_t delta;
 	int self;
-	int cnt = MIN(max, THINKOS_THREADS_MAX + 1);
+	int cnt = MIN((int)max, THINKOS_THREADS_MAX + 1);
 
 	self = thinkos_rt.active;
 	cyccnt = CM3_DWT->cyccnt;
